Almost final; Added admin panel; Attempted contact us form; A lot of dynamic blade/html tweaks;

+++ app/Http/Controllers/UserController.php
@@ -46,8 +46,8 @@ class UserController extends Controller
         $user = User::findOrFail($id);
 
         $this->validate($request, [
-            'nickname' => 'required|min:5|max:10|unique:users,nickname,' . $user->id,
-            'email' => 'unique:users,email,' . $user->id
+            'nickname' => 'min:5|max:10|unique:users,nickname,' . $user->id,
+            'email' => 'required|unique:users,email,' . $user->id
         ]);
 
         $user->update($request->all());
@@ -60,8 +60,9 @@ class UserController extends Controller
         if ($request->hasFile('avatar')) {
             $user = Auth::user();
             $filename = $user->avatar;
-            Storage::delete('/public/uploads/avatars/' . $filename);
-
+            if ($filename != 'Default.png') {
+                Storage::delete('/public/uploads/avatars/' . $filename);
+            }
             $avatar = $request->file('avatar');
 
             $circle_mask = Image::canvas(500, 500, 'rgba(0, 0, 0, 0)');
@@ -73,7 +74,6 @@ class UserController extends Controller
             Image::make($avatar)->fit(250)->encode('png')->mask($circle_mask)
                 ->insert(public_path('/img/watermark_new.png'), 'center')->save(public_path('/uploads/avatars/' . $filename));
 
-//            $user = Auth::user();
             $user->avatar = $filename;
             $user->save();
 
@@ -104,10 +104,10 @@ class UserController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
+
     public function destroy($id)
     {
-        //Didn't add the delete profile functionality yet
-        $this->middleware('auth');
+        //User delete not implemented as it's not neccessary
         $user = User::find($id);
         $user->delete();
         return redirect()->route('profile.edit');
